Rename to dag_algo
